events1

+++ app/seeds/events.py
@@ -44,7 +44,7 @@ def seed_events():
     db.session.commit()
 # add events
 def undo_events():
-    if environment =="Production":
+    if environment =="Production" and SCHEMA:
         db.session.execute(f"TRUNCATE table {SCHEMA}.events RESTART IDENTITY CASCADE;")
     else:
         db.session.execute(text("DELETE FROM events"))
